Unidad 5/requirements.txt
requests
django
djangorestframework
daphne
channels
channels_redis
selenium
python-dotenv
mysqlclient
